only programa service test

+++ backend/sppg/src/test/java/br/ufma/sppg/model/ProducaoTest.java
@@ -8,28 +8,26 @@ import br.ufma.sppg.repo.ProducaoRepository;
 import java.text.ParseException;
 import java.util.Optional;
 
-
-
 @SpringBootTest
 public class ProducaoTest {
-    
+
     @Autowired
     ProducaoRepository prod;
 
     @Test
     public void deveSalvarProducao() throws ParseException {
-        //cenário
-        
+        // cenário
+
         Producao novaProducao = Producao.builder().id(1).tipo("A").ano(2022).issnOuSigla("ABC")
-                                .nomeLocal("Ufma").titulo("Classificação de Cancer de mama")
-                                .qualis("qualisemxlm").percentileOuH5(2).qtdGrad(10)
-                                .qtdMestrado(20).qtdDoutorado(10).build();                                  
-        //acao
+                .nomeLocal("Ufma").titulo("Classificação de Cancer de mama")
+                .qualis("qualisemxlm").percentileOuH5(Float.valueOf(2)).qtdGrad(10)
+                .qtdMestrado(20).qtdDoutorado(10).build();
+        // acao
         Producao prodSalvo = prod.save(novaProducao);
 
-        //teste
+        // teste
         Assertions.assertNotNull(prodSalvo);
-        Assertions.assertEquals(novaProducao,prodSalvo);
+        Assertions.assertEquals(novaProducao, prodSalvo);
         Assertions.assertEquals(novaProducao.getId(), prodSalvo.getId());
         Assertions.assertEquals(novaProducao.getAno(), prodSalvo.getAno());
         Assertions.assertEquals(novaProducao.getPercentileOuH5(), prodSalvo.getPercentileOuH5());
@@ -40,28 +38,25 @@ public class ProducaoTest {
         Assertions.assertEquals(novaProducao.getQtdMestrado(), prodSalvo.getQtdMestrado());
         Assertions.assertEquals(novaProducao.getTipo(), prodSalvo.getTipo());
         Assertions.assertEquals(novaProducao.getTitulo(), prodSalvo.getTitulo());
-        Assertions.assertEquals(novaProducao.getIssnOuSigla(), prodSalvo.getIssnOuSigla());     
+        Assertions.assertEquals(novaProducao.getIssnOuSigla(), prodSalvo.getIssnOuSigla());
 
     }
 
-
-    
-    public void deveAtualizarEstatisticasProducao() throws ParseException{
-        //cenário 
+    public void deveAtualizarEstatisticasProducao() throws ParseException {
+        // cenário
         Producao novaProducao = Producao.builder().id(1).tipo("A").issnOuSigla("ABC")
-                                .nomeLocal("Ufma").titulo("Classificação de Cancer de mama")
-                                .qualis("qualisemxlm").percentileOuH5(2).qtdGrad(10)
-                                .qtdMestrado(20).qtdDoutorado(10).build();    
-                                   
-        
-        //acao
-        //original
+                .nomeLocal("Ufma").titulo("Classificação de Cancer de mama")
+                .qualis("qualisemxlm").percentileOuH5(Float.valueOf(2)).qtdGrad(10)
+                .qtdMestrado(20).qtdDoutorado(10).build();
+
+        // acao
+        // original
         Producao prodSalvo = prod.save(novaProducao);
-        //verificando se salvou
+        // verificando se salvou
         Assertions.assertNotNull(prodSalvo.getId());
         Assertions.assertNotNull(prodSalvo);
 
-        //atualizando componentes da producao
+        // atualizando componentes da producao
         prodSalvo.setAno(2020);
         prodSalvo.setTipo("A");
         prodSalvo.setQtdDoutorado(10);
@@ -69,23 +64,22 @@ public class ProducaoTest {
 
         // Salvar a producao atualizada
         prod.save(prodSalvo);
-    
-        //verificacao
-        //procurando a producao pelo id no repositório
-        Integer id = prodSalvo.getId(); 
-        Optional <Producao> temp = prod.findById(id);       
+
+        // verificacao
+        // procurando a producao pelo id no repositório
+        Integer id = prodSalvo.getId();
+        Optional<Producao> temp = prod.findById(id);
         Assertions.assertTrue(temp.isPresent());
 
-        // atribuindo a producao encontrada a uma nova variável 
-        Producao prodAtualizado =  temp.get();
+        // atribuindo a producao encontrada a uma nova variável
+        Producao prodAtualizado = temp.get();
 
-        Assertions.assertNotEquals(novaProducao.getAno(),prodAtualizado.getAno());
-        Assertions.assertNotEquals(novaProducao.getTipo(),prodAtualizado.getTipo());
-        Assertions.assertNotEquals(novaProducao.getTitulo(),prodAtualizado.getTitulo());
-        Assertions.assertNotEquals(novaProducao.getQtdDoutorado(),prodAtualizado.getQtdDoutorado());
+        Assertions.assertNotEquals(novaProducao.getAno(), prodAtualizado.getAno());
+        Assertions.assertNotEquals(novaProducao.getTipo(), prodAtualizado.getTipo());
+        Assertions.assertNotEquals(novaProducao.getTitulo(), prodAtualizado.getTitulo());
+        Assertions.assertNotEquals(novaProducao.getQtdDoutorado(), prodAtualizado.getQtdDoutorado());
         Assertions.assertNotEquals(novaProducao.getQualis(), prodAtualizado.getQualis());
 
     }
-    
 
 }
